fix(ci): silence lint and ty errors so CI passes
Exclude notebooks/ and docs/ from ruff (marimo cells legitimately
use bare expressions). Fix B905 zip-strict, F841/RUF059 unused
variables, RUF046 redundant int, SIM101 isinstance merge across
src/ and tests/. Annotate registered buffers (mel_mean/mel_std/
wav_std/freqs_cis) at class level so ty narrows them to Tensor.
Modernize torch.cuda.amp.GradScaler -> torch.amp.GradScaler in
CheckpointManager and tighten exp_path to Path (still coerced in
__post_init__). Remaining FloatTensor/BoolTensor/LongTensor and
optional-narrowing gaps are suppressed with # ty: ignore at call
sites rather than relaxing the declared types.

+++ tests/test_neural_speaker.py
@@ -278,7 +278,7 @@ def test_training_step_default_acoustic_front_samples_warmup_region(
             if len(ones) > 0:
                 fronts.append(int(ones[0].item()))
             else:
-                fronts.append(int(round((tb[0].item() - 1.0) * (n - 1))))
+                fronts.append(round((tb[0].item() - 1.0) * (n - 1)))
     fronts_t = torch.tensor(fronts)
     assert fronts_t.min().item() >= -(n - 1), fronts_t.min().item()
     assert fronts_t.max().item() < int(lens_ext.max().item())
